Update patch set 2
Patch Set 2:

This is a workaround for the case when mirror is not available due to update. I've opened a bug[1] for this and found a workaround[2]

I am not sure if this is the right way to do, maybe there will be better suggestions or maybe this issue is too minor and rare to be fixed.


[1] https://launchpad.net/bugs/1603001
[2]:http://askubuntu.com/questions/41605/trouble-downloading-packages-list-due-to-a-hash-sum-mismatch-error

Patch-set: 2
Reviewer: Gerrit User 14807 <14807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0