nova: n-sproxy config consistency
This change makes the configuration of the serial_console section in the
nova.conf file consistent to that like the other console sections (VNC/SPICE).
That means, if the n-sproxy service is not enabled, disable the feature in
the nova.conf file. This should avoid confusion for people who expect the
feature to work after enabling the necessary service.

Change-Id: I5977b9a275e1451fa3baf37ae6e1130371249fef