Support patching projects/plugins
Add a patching mechanism right after cloning the
projects and fetching plugins, in order to allow the
user to supply patches. User will place his patches
under a "patches" directory under the devstack root
directory. The patch name will start with the name of
the project/plugin, and end with ".patch", and between
the user can put a string to identify the patch.

Change-Id: Ie28e1f8dda4ea138f08a3a8e3962bdfbe7ec2b06
Closes-Bug: #1606625