Update patch set 2
Patch Set 2: Code-Review-1

I realise this is relatively minor code, but I would not like to see this approach in devstack core.  The reason is that next thing you know this will be used in the gate and we'll be making it even harder to replicate gate failures than it already is by pulling in code changes outside our usual mechanisms.  The idea that "we're running $PROJECT, but with the XYZ company patch" where that patch is built & maintained outside the usual OpenStack review mechanism is not appealing.

I also realise nothing stops a plugin doing this anyway, but devstack does have to make decisions about what we support.

Our way of bringing in patches is really via the review mechanism [1].  With depends-on you can get cross-project dependencies.  if something isn't quite ready you can always wip it.  It's kind of the core thing about openstack that we all use gerrit and collaborate on changes.

I'm willing to listen to the use-case for this and maybe it's convincing ... but that's my first thought here.

[1] I'm not sure if you know this -- gerrit does its best to keep it hidden in the "Download" pull-down -- but you can pull this change into a tree with "git pull https://git.openstack.org/openstack-dev/devstack refs/changes/78/347478/2"

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1