Update patch set 2
Patch Set 2:

> I realise this is relatively minor code, but I would not like to
 > see this approach in devstack core.  The reason is that next thing
 > you know this will be used in the gate and we'll be making it even
 > harder to replicate gate failures than it already is by pulling in
 > code changes outside our usual mechanisms.  The idea that "we're
 > running $PROJECT, but with the XYZ company patch" where that patch
 > is built & maintained outside the usual OpenStack review mechanism
 > is not appealing.
 > 
 > I also realise nothing stops a plugin doing this anyway, but
 > devstack does have to make decisions about what we support.
 > 
 > Our way of bringing in patches is really via the review mechanism
 > [1].  With depends-on you can get cross-project dependencies.  if
 > something isn't quite ready you can always wip it.  It's kind of
 > the core thing about openstack that we all use gerrit and
 > collaborate on changes.
 > 
 > I'm willing to listen to the use-case for this and maybe it's
 > convincing ... but that's my first thought here.
 > 
 > [1] I'm not sure if you know this -- gerrit does its best to keep
 > it hidden in the "Download" pull-down -- but you can pull this
 > change into a tree with "git pull https://git.openstack.org/openstack-dev/devstack
 > refs/changes/78/347478/2"

Hi Ian,

While I agree that patches/fixes should always go to gerrit, the process is sometimes too long. Some projects don't have many reviewers, and sometimes it just takes time until a fix is committed, and it is obviously impossible to work with devstack if you have to wait for a fix. I do believe, that everyone that uses devstack to do something more than trivial, has something similar to this commit.
If we want devstack users to be satisfied, we must solve the problem of waiting for commits.
In our company, waiting for a fix-commit is simply not an option. So we do whatever we can, in order to push forward the fix-commit, but we can't stop testing/developing until it is committed.
We have already been using this patch_plugins method for a while now, and we just wanted to share it with the community, as we are certain it will benefit from it.
I totally agree about the gating comment - If gates run with different commits, it will definitely make it harder to reproduce bugs, but again - If the gate cannot run because it is waiting for a fix, and the fix-commit is taking time, what is the point of the gate? As you said - whoever is running the gate will do it anyway if they want to, so better standardize it. Maybe we can even find a way to have a gate output whatever patches it was running with (if it was running with).

That said, I'm always open to other suggestions, especially if they are better :-)

Patch-set: 2
Reviewer: Gerrit User 22305 <22305@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0