Update patch set 2
Patch Set 2:

> it is obviously impossible to work with devstack if you have to wait for a fix.

I'm not sure I would agree with this; fundamentally we are using source from git and thus the thing it does really well is manage changes.

It may not be totally obvious -- git links are hard to find on gerrit, maybe we don't document it well in devstack.  it's still not quite clear to me exactly what workflow is holding you up and requiring patches.  it might be we just need better documentation to help expose how to best use the tools we have

Patch-set: 2