DNM: Debug state of image dir
I have a feeling I need to land a patch first that adds the qcow image
to the IMAGE_URLs list so that nodepool will start preseeding it on
the images prior to running the test. But I thought it would properly
fallback and actually hit the internets and download trhe image. This
patch just is going to print the state of the image dir so I can tell
what's going on.

Change-Id: I52caaa260541997481d0dfd0cf04539d54858083