Always install apache and mod_wsgi
Most services now support being deploying in apache with mod_wsgi in
some form or another. In trying to standardize how we deploy SSL and
non-SSL sites in apache it is easier to simply always install apache and
mod_wsgi initially. In the unlikely event nothing is configured to use
them they are not a problem to have on the system.

Change-Id: I521cd7080d9f0d7ac74ad550780e6506a88124ed