Setup apache snippets structure
To make all devstack services work in a namespaced fashion we need for
them all to be able to individually add apache configuration that is
read as part of the same virtual host.

The way different services treat apache makes it difficult to share,
currently horizon simply creates and owns the port 80 virtual host,
keystone needs to add things to the port 80 host and create the 443
vhost to make SSL work.

This arrangement is not just a keystone/horizon problem and will apply
to all services.

Create a vhost snippet arrangement. A "http" site is added to apache
that loads all apache config snippets from "http-enabled". This
configuration will in future patches allow a similar "ssl-enabled"
structure for SSL endpoints.

Now everything that is added to the http-enabled snippets exists in the
same vhost, yet each service can control its own apache config.

Change-Id: I1aca5d0fded41f5bb376278878e8bb84a18fe695