Update patch set 1
Patch Set 1:

> If there is no route to $ROUTER_GW_IP that tells me that something
 > is misconfigured.

Thank you for your reviewing. 

This is not configuration problem. "RTNETLINK answers: Network is unreachable" happened when public interface not configure an IP address.

I think that add a route to private network is unnecessary,  maybe this is convenient in a development environment, but this shouldn't cause stack.sh failed when failed to add route.

 > If there is no route to $ROUTER_GW_IP that tells me that something
 > is misconfigured.

Thank you for your reviewing. 

This is not configuration problem. "RTNETLINK answers: Network is unreachable" happened when public interface not configure an IP address.

I think that add a route to private network is unnecessary,  maybe this is convenient in a development environment, but this shouldn't cause stack.sh failed when failed to add route.

Patch-set: 1
Reviewer: Gerrit User 21778 <21778@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0