Update patch set 1
Patch Set 1:

@Ian, I have to sort out https://review.openstack.org/#/c/336949/ - I'll probably just change it to add a debug log message that files were provided for injection but it's disabled. It's tricky because the files could be served via config drive (which I can check from the compute) or the metadata service, which I can't check from the compute, at least not in an obvious way I can think of - or anyone else has suggested.

It's only related to Id0c521f6f624367bd497463c8c2d99488548fcff in the fact it's about file injection and trying to clean this all up.

And yes, the neutron full ssh job was broken this whole time but no one really noticed because it was/is a non-voting job. I'm working on making it voting which is how I stumbled onto this.

Patch-set: 1