Add fixed_key in cinder.conf
When Cinder creates an encrypted volume from image, it accesses
the key manager to get the key, encryptes the data, and writes to
volume. This feature is introduced in
https://review.openstack.org/#/c/216567/.

By default, it uses ConfKeyManager by default. The key manager needs
the parameter fixed_key set in config file. This key manager is just
for test, not for customer usage.

Change-Id: I1265cc2af2acb1d950c7a944e607a9707d2f69e6