Update patch set 11
Patch Set 11: Code-Review-1

So the code looks good to me, but my fundamental issue before with this patch was actually stemming from a lack of explanation of what is going on here. Matt had to dig into the cinder code and the tempest code to figure out what was happening. This needs inline comments explaining why that loop to split things is there. The commit message should also explain the difference between cinder and tempest config files.

Patch-set: 11
Label: Code-Review=-1