Copy tempest config to /etc/tempest
When using the recommended tempest init && tempest run workflow we want
to prepopulate the tempest workspaces created by running tempest init
with the devstack generated tempest config file. This way it's a
seamless experience for using the new workflow. It also will enable us
to start the migration of devstack-gate over to using the new workflow.
This commit enables that by copying the tempest config file in to
/etc/tempest. /etc/tempest is one of the locations tempest init will
use to preseed the local workspace etc/ directory.

Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>
Partially-Implements: bp tempest-run-cmd
Change-Id: Ida35a27fc441847351084be033e35e9a04042195