Update patch set 1
Patch Set 1:

The idea is to provide consistency especially from a CI perspective (especially 3rd party stuff).  We do things like grep output which needs to work everywhere.

i agree that overriding the client settings in stackrc is too much.  if we were to support running stack.sh in different locales with a flag or something, i think we should consider that in a separate change.  but i think it would be too backwards-incompatible to just remove it, since it's likely to break people's CI

Patch-set: 1