Update patch set 2
Patch Set 2: Code-Review-1

I'm a bit concerned in just making this change to get past testing, where are the release notes associated with the fact that we need a shared lock space? And if so, it feels like we need to revisit this fact with oslo.concurrency and not make users provide the right lock path themselves or get breakage.

Patch-set: 2