Update patch set 1
Patch Set 1: Code-Review-1

I worry about doing it this way, it means our test configuration is dependent on the neutron api to work properly. If neutron lands a bug in their extension list code that returns the wrong set (which is too small) this means we silently end up skipping stuff. For the gate this is not something I think we want to rely on.

Patch-set: 1