Update patch set 4
Patch Set 4: Code-Review-1

this might get us past this round of the mirror updates, but if they are that inconsistent, there is likely to be another issue later when trying to actually install these packages.

And this makes the code much harder to understand. My preference is to keep this out, and realize that this is a problem which happens from time to time.

Patch-set: 4
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1