Update patch set 2
Patch Set 2:

I think this would make it unnecessarily complex. The config option is designed to be equivalent to ENABLED_SERVICES where one can pick and choose what to enable (for SSL in this case). Adding and deleting services could produce unexpected results for those that don't know what services are enabled for SSL by default.

The reason this is mostly necessary is the various projects fragility when it comes to SSL support. For example, when this patch was submitted cinder was broken. The only workaround was to directly modify stack.sh which wasn't so nice.

Ideally we get to the point of stability and gate jobs to require SSL (and proxy) support to work before accepting patches but we aren't there yet. This is a stop-gap to allow people to experiment with SSL mode in an understandable way without making changes to code.

But even then its likely that only a blessed subset of services will be tested regularly. It may be desirable to enable services for different tests simply by updating the list in configuration.

Patch-set: 2
Reviewer: Gerrit User 7662 <7662@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0