Update patch set 2
Patch Set 2:

Sorry for the delayed response, I'm actively wanting to push this stuff but am traveling at the moment. So, I can appreciate some good old fashion bike shedding about locking as much as the next dev... but I do think we could maybe take a multi-pronged approach to tackle this.


A couple of points I think are important to keep in mind for this:

* Its only really an issue if c-vol and n-cpu are on the same node, this doesn't seem super common in "real" deployments  (hence why no one seemed to notice this as an issue in all the years it has been working this way...)

* Even if we change the way the locking works, it will take time and more than likely will not make it into N. We'll be testing and using DevStack with this code for a while :(


What I would propose is that we go ahead with this bandaid in DevStack, and make sure that we raise up the issue to Cinder, Nova, and the Doc writers so that its can be correctly documented somewhere. In parallel to this we can start a discussion on a longer term solution for this that we can all agree on and get community support for.


Does that sound reasonable?

Patch-set: 2
Reviewer: Gerrit User 12924 <12924@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0