Update patch set 1
Patch Set 1: Code-Review+2

So I confirmed the extension list on master for the gate isn't changing:

http://logs.openstack.org/88/352588/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/e4611a4/logs/tempest_conf.txt.gz

and we still correctly set it to 'all' in teh gate. I would have liked to see this have the depends on for the devstack-gate sanity check patch. But I doubt we'll slip and enable this during that window

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2