Update patch set 2
Patch Set 2: Code-Review-2

not wanting to put a work around in devstack to corruption by default behavior in cinder/nova is not bikeshedding.

I also think this is something that the cinder/nova teams are probably not fully aware is an issue.

So, please start with landing the release note in both of those projects so it's clear that the core teams on both understand what's being asked of operators here to not corrupt their data by default, with no warning.

After the release notes are in, this is game to land in devstack.

Patch-set: 2
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2