Update patch set 2
Patch Set 2:

As requested:

https://review.openstack.org/354501

https://review.openstack.org/354502


As an aside, I kinda disagree with the tactic of holding a bug fix hostage like this... IMO (heavy emphasis on the "my opinion"), this change needs to happen either way, so blocking it just kind of.. what? Encouragement? Luckily as an active community member and (still IMO) a pretty reasonable human I'm game for documenting broken stuff, but had I fucked off and not followed through we would still just have a broken config in Devstack. What good is that?

I'll admit my comment earlier on bike-shedding was a tad uncalled for, so I apologize if it sounded like I was dismissing the need to fix the underlying issue as unworthy of pursuit. What I meant (mostly as a joke) was that a thread on openstack-dev to delve into "the larger question of what is that API, how does it work, what is the right description of that API." would surely not be quick, on topic, or be solving the volume attachment problems of current nova and cinder needing this config option set. Again though, thats just like, my opinion.

Patch-set: 2