Fix incorrect cinder start order
In Mitaka's Cinder Upgrade Notes [1] it is mentioned that the
cinder-backup service should be the last to be upgraded and restarted:

"As cinder-backup was strongly reworked in this release, the recommended
upgrade order when executing live (rolling) upgrade is
c-api->c-sch->c-vol->c-bak."

But devstack starts Cinder Backup service before Cinder Volume, which
can lead to problems during grenade jobs [2] where we see this error:

"ServiceTooOld: One of cinder-volume services is too old to accept such
request. Are you running mixed Liberty-Mitaka cinder-volumes?"

This patch changes the order so the cinder volume service is started
before the backup service.

[1]: http://docs.openstack.org/releasenotes/cinder/mitaka.html#id6
[2]: http://logs.openstack.org/58/350758/1/check/gate-grenade-dsvm-designate-powerdns/90b0f2d/logs/new/screen-c-bak.txt.gz?level=ERROR

Change-Id: Idca0abc45ca314c4319c39f9a8d59ebeac7a720e
Closes-Bug: #1612781