Allow for turning on Cinder CG API's
There is a new flag to enable them CINDER_ENABLE_CG_APIS which
defaults to False. If set to True it will update the policy.

Modify policy_add helper to not only add but modify any
existing policy that might be there. We then use policy_add
in lib/cinder to set ConsistencyGroup API's as needed.

policy_add was added to devstack for grenade in [1], however
grenade stopped using it in [2].  It was only ever added
for cinder in [3]. Codesearch shows no other users, so
this modification in behavior should not affect anyone.

[1] https://review.openstack.org/#/c/44546/
[2] https://review.openstack.org/#/c/171601/
[3] https://review.openstack.org/#/c/42971/

Change-Id: I26313a0426537c9bf45ff8ea3df73c730ac8b1e3