Add RabbitMQ expiration policies
This fix services with interruption signal and leave the target queues
open. The number of queues keep growing by repeat above action.
This patch add a expiration for queues. allow queues to be close as
long as they didn't been used for a long period of time(1 hour).
This fix must solve problem with growing number of RabbitMQ queues.
Closes-Bug: #1599104

Change-Id: I2d4b73f217c1d7b25a3353d4c4e7c54642538b40