Neutron plugin cannot create multiple subnets
The Neutron plugin in devstack has hard-coded values for subnet
creation.  In Trove we need to create a network/subnets for the alt_demo
tenant, however that is not possible using the neutron plugin methods.
Passing in some of the variables to the neutron_create_private_subnet_*
methods allows different subnet creation without duplicating the
Neutron plugin code.

The '_' was removed from the two functions to indicate that they can
now be called from outside the Neutron plugin.

Change-Id: I865009d534412b8bdd0973b7af5adbb5fb1047ba