Update patch set 2
Patch Set 2:

(2 comments)

Hi Devananda, thanks for reviewing!

The main point here is to ensure that these features are disabled when ironic is running and do not rely on config defaults (that may change).

Most of these features are enabled by default (i.e. network port_admin_state_change) and the related tests are skipped when ironic is up. Now we need to ensure this feature is disabled for ironic to run.

Patch-set: 2
Label: Verified=0