Install python-openstackclient before plugins
Everything except for OSC installs before plugins, then plugins run.
That's great, because if a devstack plugin needs to do something
explicit with the software installed, it can. But then OSC gets
installed and can step on the install that was done by the plugin.

Just move it to before the plugin, so that plugins can have the final
say on versions of things installed.

Change-Id: I2c5658945fda5a66aa26fc4ac111f109ba79c5a6