Docs: Pin IP_VERSION to 4 for provider networks
L3 Service functions are always being sourced [1]
so the l3 agent sets the IP_VERSION=4+6 even if
l3 is not activated in local.conf. Due to this and
that the examples do not specify an IPV6 range and gateway,
for provider networks. Devstack terminates with
"IPV6_PROVIDER_FIXED_RANGE has not been set,
but Q_USE_PROVIDERNET_FOR_PUBLIC is true and IP_VERSION includes 6"

[1] https://github.com/openstack-dev/devstack/blob/master/lib/neutron-legacy#L268

Change-Id: Ib4979e725770ec473abb62af4a4f3444c164dbea