L3 correct bridge for external from OVS_PHYSICAL_BRIDGE to PUBLIC_BRIDGE
by default "OVS_PHYSICAL_BRIDGE" and "PUBLIC_BRIDGE" are set to "br-ex"
that mean we use 1 bridge in OpenStack.

This patch fix l3 to allow a case when external and internal networking
are different bridges and different interfaces.

The patch also retain the old behaiver because both paramters are set to
same bridge name by defualt.

controller local.conf example:
http://paste.openstack.org/show/584396/

Change-Id: I82b2a437b9d262ddaf092c5f1ecbcd125ed9fe8b