Update patch set 4
Patch Set 4:

@Sean:

for example, the vagrantfile that we use for designate development mounts some repos into the /opt/stack folder via nfs. [1]

depending on how many of these repos are present, the time it takes for the sudo chown stack:stack /opt/stack to run is anything from 3 mins to 20 mins.

The "is_nfs_directory" check is only ran against /opt/stack - it does not look at folders within it. in this case, we could set CHOWN_EXCLUDED_DIRECTORIES to be "/opt/stack" - as everything is either nfs, or cloned by the stack user, and does not need the chown ran on it.

I am also experimenting with passthrough disk mounting from libvirt, and in that case we definitely do not want to run the chown - it could have some serious side effects.


1 - http://git.openstack.org/cgit/openstack/designate/tree/contrib/vagrant/Vagrantfile

Patch-set: 4
Label: Verified=0