Update patch set 2
Patch Set 2:

PHYSICAL_INTERFACE is also a parameter that used in openstack/neutron. the terms here are not clear public/physical can be sperated interface or the same. this is what I'm try to achieve here. by referencing PUBLIC_INTERFACE to OVS_PHYSICAL_BRIDGE. As I also wanted to clearly 
in this commit https://review.openstack.org/#/c/360490/

Patch-set: 2