Update patch set 1
Patch Set 1: Code-Review-1

upstream we cache this for similar reasons of not hitting the mirrors.  i'll admit that it's not obvious, but the idea is that if the file is there we don't try to download it.  in infra we download it during image build with [1]

So i guess this doesn't exactly hurt ... but i'd like to see a comment explaining the situation and that probably your setup infrastructure should put this in to be found

[1] http://git.openstack.org/cgit/openstack-infra/project-config/tree/nodepool/elements/cache-devstack/source-repository-pip

Patch-set: 1
Label: Code-Review=-1