Update patch set 1
Patch Set 1:

you might be hitting [1] -- if devstack did download the file for you in the first place, it will check for updates if it sees the ".download" file.  this was because people would leave devstack around for 6 months and then report bugs that were due to old pip versions.

but on a fresh install, a copy of get-pip.py in files should stop the download ... this is what upstream is doing thousands of times a day :)

[1] https://git.openstack.org/cgit/openstack-dev/devstack/tree/tools/install_pip.sh#n56

Patch-set: 1