Move RABBIT_USERID and RABBIT_HOST to lib/rpc_backend
This moves setting of RABBIT_USERID from stack.sh to lib/rpc_backend
so it may be used in grenade runs, which don't have the defaulted
value from stack.sh. The RABBIT_USERID is needed in order to call
get_transport_url in lib/rpc_backend.

RABBIT_HOST is also moved as a related variable.

Change-Id: I6f211e9102f79418f9f94a15784f91c4150ab8a7