Update patch set 1
Patch Set 1:

Got it, thanks!

But for manual testing on a fresh system I still need 4 steps to get devstack running: 
- checkout devstack
- add local.conf
- manually download get-pip.py
- stack.sh

If I could configure an alternative get-pip location via the local.conf, I could skip the manual download. That's why I started this patch.

I leave it up to you - if you don't want this configuration in devstack, I'll abandon it and go with the 4 steps.

In general I think it's helpful to keep URLs configurable (with a sane default) instead of hardcodeding them.

Thanks for your feedback!

Patch-set: 1