Update patch set 1
Patch Set 1:

I'm not really that opposed to this, it's trivial, but if we put it in I'd like to see it with expanded documentation covering what's going on.

If you're really doing this with any frequency, it's probably worth making VM images with project-config:tools/build_image.sh ... they will be pretty close to what we use upstream and have all the caching done so you don't have to do unreliable downloads.

Patch-set: 1