Default nova cpu allocation ratio to 4.0
The default cpu allocation ratio in nova is 0.0, which means no
overcommit. This is good for production, but not really for developers
and the gate. Since we don't have CoreFilter enabled by default, people
don't notice that the scheduler is technically overcommitting our
cpus, against the advice of this config.

Since this value is used in the brave new world order of resource
providers, we're getting warnings (which will be errors later) because
the non-resource-provider scheduler is overcommitting things. By
defaulting this to something sane for developers and the gate, we
avoid these warnings, and later errors.

Change-Id: If222f7f12fcddc3dc03ebf78b546fed51b6e8a81