Adds neutron_lbaas.conf to q-svc command line
When q-lbaasv2 is enabled in local.conf, the neutron-lbaas plugin.sh script
creates a new /etc/neutron/neutron_lbaas.conf file with some config parameters.

Under several circumstances, some of the options in this file are needed by
other neutron daemons, such as the q-svc service.

This patch adds a check to examine if q-lbaasv2 is enabled (excluding octavia),
And on such cases, it makes sure that the command-line for the q-svc service
will also include '--config-file /etc/neutron/neutron_lbaas.conf' so that these
configuration parameters are pulled in for that service.

Change-Id: Ia57cc4acd7bca2711b6f0449dd3443a62ddd9c95
Closes-Bug: #1619466
Related-Bug: #1565511