Update patch set 1
Patch Set 1:

I think the neutron only solution is better, though we should solve the issue of not being able to append drivers to the config option. We may need to do some generic work in devstack to expose a function to append a driver, and then reuse it from neutron tree to enable the driver in gate.

Patch-set: 1