nova: stop setting deprecated use_usb_tablet option
The use_usb_tablet option is replaced by pointer_model
option. The default behavior is different though, so we have
to set pointer_model to the empty string because it's default
of None will trigger a fallback to the use_usb_tablet option
which defaults to True, which would change the behavior here.

We can remove this and rely on the default value of None once
the use_usb_tablet option is removed from nova in Ocata.

Change-Id: Ic2a49f88df988c6404c1c72e9ee28a487e4f7908