Update patch set 2
Patch Set 2:

Is there anything that nova-lxd could do to make this easier? I see the patch is quite small, but it's a little odd to add this to devstack when the lxd driver is not (yet) in nova proper. Maybe something we could add to our devstack plugin?

Patch-set: 2