force en_US.UTF-8 during devstack runs
devstack was defaulting to C locale to avoid parsing issues with a
number of bash commands. However, as it starts all the services under
that locale it actually means we're not really running in a unicoded
enabled service environment. That masks a number of issues.

We should stop doing that, and force that we're in a known utf8
environment for command parsing.

Change-Id: Ie121a8ebf8d5d8ddd75eb1688be7a91893a8eba6