Load br_netfilter module if present
This module is now needed to get access to firewall for bridged frames,
as well as to /proc/ settings that are used by Neutron to enforce
firewall for all bridges.

Since nova-network does not pass -m physdev to any of its iptables
rules as Neutron does, the module is not automatically loaded, leaving
the firewall ineffective. For Neutron, the module loading is triggered
by the first -m physdev rule applied, so in general case there is no
visible breakage.

http://ebtables.netfilter.org/documentation/bridge-nf.html

Change-Id: I01cc6fe9198730a629a39349551ca3494dd4ac3a
Closes-Bug: #1611895
Closes-Bug: #1622914