Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

So I'm -1 on this because we shouldn't rely on the neutron api to make tempest configuration decisions for gating. I'm fine with using it outside the gate though. We should have enough information from devstack to know it's setting the variable in the ml2.conf, so we just need to find the best way to get that info into lib/tempest

Patch-set: 2
Label: Code-Review=-1