Allow for chown & chmod operations on DEST to be toggled
If you are using libvirt + NFS, this step can take quite a long time,
or if you are using libvirt + passthrough filesystems, you really
don't want this to run this.

This is due to the nfs check only checking if the root dir is
nfs. some setups have a local /opt/stack with other
directories nfs mounted into /opt/stack, and as such the
"is_nfs_directory" test allows the chmod / chown to run.

This allows users to disable these operations by setting

  CHOWN_DEST & CHMOD_DEST to False

Change-Id: Ia10abd70e712d42ff78117bac932685ef119fbbd