Update patch set 5
Patch Set 5:

Sean, we need that for nova-network. Nova folks already said they won't handle it in their code.

As for Neutron, we could in theory do it by allowing modprobe to be executed by the agent. You think it's a legit thing to do? Nothing that would scare you about exposing kernel injection to the agent?

Patch-set: 5