Ironic: Set default images for UEFI deployment
This patch is adding support for setting default UEFI images in case
Ironic is configured to deploy nodes in UEFI mode.

It's important to note that the current version of CirrOS 3.4.0 was
released in May 2015 and does not support UEFI. Therefore, we need to
use the daily built images after 21/07/2016 which is when the UEFI
supported landed [0].

For the moment, when UEFI is requested this patch is downloading the
images from 22/07/2016 until a new release of CirrOS is available.

[0] https://code.launchpad.net/~hrw/cirros/add-uefi-grub-support/+merge/299185

Change-Id: Iedbbec33d656c843697403eef96fd96652307df4
Depends-On: I4e9247c6a91c707db08247a8ac04b1f146cd3714
Partial-Bug:#1625616