Adds keystone security compliance settings
The PCI-DSS feature has been introduced during the Newton
release and its settings are disabled by default. This
patch adds the possibility to enable some of them during
DevStack setup.

This change is intended to enable the possibility to
create functional tests for the feature for keystone's
tempest plugin job.

Change-Id: If6b5eb3e3cbc43eb241c94d18af80ad50be08772