Update patch set 1
Patch Set 1: Code-Review+2

Yes we're currently writing that conf variable to the file even if neutron is enabled:

http://logs.openstack.org/49/373249/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/b0159bf/logs/tempest_conf.txt.gz

and I confirmed this fixes the intent of that if check:

http://logs.openstack.org/11/374311/1/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/ca28628/logs/tempest_conf.txt.gz

So this does fix the logic there. It does raise the larger question of why we're blocking it on neutron. For the default devstack tempest configuration this variable doesn't break anything (since we configure tempest to use dynamic credentials) and all it does it enable a couple extra tests that do a server list with a filter on networks. But we can address that in a follow on change

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2