Update patch set 2
Patch Set 2:

Is the confusion around the hardcoded number '3'?  This isn't the number of elements in the list of virt drivers, but rather the total number of images identified to download.  Each driver could require an arbitrary number of images and request they be cached, but the impact of reducing the list of drivers in the project-config change is that the original hardcoded check of 5 images is more than will be cached by default for Infra's runs

Patch-set: 2