Update patch set 2
Patch Set 2:

If you use neutron, you can't rely on that key. You need to dynamically get the project network, which can be dynamically created. Unfortunately it seems that there is no public tempest API for this, but if your base test class still relies on the non-stable  BaseTestCase class, you can use get_tenant_network and try to work on tempest to have a public version of that method/function.
In other words: if you were that key for the private network in your neutron tests, your tests results were potentially unreliable.

Yes, the logic was changed, but now it's doing what was intended: not set the key when neutron is used. The previous bash logic was broken, as you can easily check by comparing what that if was intended to do from the comment and the commit message of the review where it was changed, and what it was really doing.

Patch-set: 2