Change default IPv6 subnetpool prefix
Using the documentation prefix as real world default for IPv6 is
bad, instead use a random prefix from fd00::/8 just as for
FIXED_RANGE_V6.

We have a /48 as random prefix in total, we use a /64 from it for
FIXED_RANGE_V6 and in order to avoid overlapping the subnet pool range
with that one, we use the next /56 after the one occupied by
FIXED_RANGE_V6. This will allow for the creation of 256 subnets out of
this pool.

Change-Id: Ied96ea4faff05c8f62dbb1ec46a6903d672c71d8