Fix api extension settings in tempest for Liberty
On stable branches, we are supposed to hard code the extension list of
what API extensions are available when we branch the project so that
test cases against new features will be skipped. However, we didn't do
this on liberty branch for neutron, cinder and swift. So copy from
kilo and added the following API extensions (not in kilo):
  network
  - dns-integration
  - flavors
  - rbac-policies

  volume
  - capabilities
  - os-snapshot-manage
  - os-snapshot-unmanage

For nova API, we are using [1] "stable API" and API extensions are
removed. So we can use "all" for "compute_api_extensions".

[1] http://docs.openstack.org/developer/nova/stable_api.html#stable-api

Change-Id: Ic4a6292390a88e732f7460cf558c479baa8771b1