Update patch set 6
Patch Set 6: Workflow-1

We realized that the need in this patch is actually dependent on the path neutron devs will choose to handle the issue. If we decide that the agent does not have business in configuring kernel, then it's needed. Otherwise, we may as well have the agent solving it with loading the module and fixing sysctl, and then it won't be necessary.

Patch-set: 6