Update patch set 14
Patch Set 14:

Ian, motivation is clear. We also have plugins, which have dependencies on other plugins. Like in sahara project we are widely using heat to provision infrastructure. It's clear that if your are using sahara in devstack, you have to enable additional plugin for heat always. From point of view of developer you always have to know all plugins which you to have to enable, but we are obviously can avoid that  by introducing requirements for plugins. I thought that is just clear.

| figuring out dependencies becomes hard.  

No, you just go through logs and it becomes clear what projects were fetched. It just adds logical encapsulation and makes stuff around enabling plugins more clear.

| i think i'd prefer to see it parsed

In such way it should be parsed? Please explain that in more details because it is absolutely unclear for me what you really mean by that. I though it is a common practice to source additional stuff because we are fetching settings, and so on...

Patch-set: 14